1° Lista avaliativa finalizada

+++ 1° Lista de exercícios - Avaliativos/Exercicio2.c
@@ -5,21 +5,24 @@
 
 int main()
 {
+
+    // Código feito por Lucas Peixoto
+
     setlocale(LC_ALL, "");
 
-    double x1, x2, y1, y2, distancia;
+    float x1, x2, y1, y2, distancia;
 
     printf("Informe o eixo x do primeiro ponto : ");
-    scanf("%lf", &x1);
+    scanf("%f", &x1);
     printf("Informe o eixo y do primeiro ponto : ");
-    scanf("%lf", &y1);
+    scanf("%f", &y1);
     printf("Informe o eixo x do segundo ponto : ");
-    scanf("%lf", &x2);
+    scanf("%f", &x2);
     printf("Informe o eixo y do segundo ponto : ");
-    scanf("%lf", &y2);
+    scanf("%f", &y2);
 
     distancia = sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
 
-    printf("A distância entre os pontos é %.4lf", distancia);
+    printf("A distância entre os pontos é %.4f", distancia);
     return 0;
 }
